Created a big amount of external objects in JsSodiumInterface.kt for Hash.kt, Kdf.kt and KeyExchange.kt files. Reason: values had these fields in dynamic type but not in JsAny type.

+++ multiplatform-crypto-libsodium-bindings/src/wasmJsMain/kotlin/com/ionspin/kotlin/crypto/box/Box.kt
@@ -14,8 +14,8 @@ actual object Box {
     actual fun keypair(): BoxKeyPair {
         val keypair = getSodium().crypto_box_keypair()
         return BoxKeyPair(
-            (keypair.publicKey as Uint8Array).toUByteArray(),
-            (keypair.privateKey as Uint8Array).toUByteArray()
+            (keypair.publicKey).toUByteArray(),
+            (keypair.privateKey).toUByteArray()
         )
     }
 
@@ -25,8 +25,8 @@ actual object Box {
     actual fun seedKeypair(seed: UByteArray): BoxKeyPair {
         val keypair = getSodium().crypto_box_seed_keypair(seed.toUInt8Array())
         return BoxKeyPair(
-            (keypair.publicKey as Uint8Array).toUByteArray(),
-            (keypair.privateKey as Uint8Array).toUByteArray()
+            (keypair.publicKey).toUByteArray(),
+            (keypair.privateKey).toUByteArray()
         )
     }
 
@@ -142,8 +142,8 @@ actual object Box {
             sendersSecretKey.toUInt8Array(),
         )
         return BoxEncryptedDataAndTag(
-            (detached.ciphertext as Uint8Array).toUByteArray(),
-            (detached.mac as Uint8Array).toUByteArray()
+            (detached.ciphertext).toUByteArray(),
+            (detached.mac).toUByteArray()
         )
     }
